+++ SoftwareUpdates/Create-PatchDeployments.ps1
@@ -149,7 +149,7 @@ Function Write-ScriptLog
                 exit 2
             }
          }
-        Write-EventLog –LogName $EventlogName –Source $Component –EntryType $Severity –EventID $EventID –Message $Message
+        Write-EventLog -LogName $EventlogName -Source $Component -EntryType $Severity -EventID $EventID -Message $Message
     }
 
     # save severity in single for cmtrace severity
@@ -184,48 +184,61 @@ Function Write-ScriptLog
 
 
 #region Rollover-Logfile
-<#
+<# 
 .Synopsis
-   Will create a new logfile if a specified file size is reached
+    Function Rollover-Logfile
+
 .DESCRIPTION
-   Will create a new logfile if a specified file size is reached
+    Will rename a logfile from ".log" to ".lo_". 
+    Old ".lo_" files will be deleted
+
+.PARAMETER MaxFileSizeKB
+    Maximum file size in KB in order to determine if a logfile needs to be rolled over or not.
+    Default value is 1024 KB.
+
 .EXAMPLE
-   Rollover-Logfile -Logfile C:\temp\logfile.log -MaxFileSizeKB 2048
-.EXAMPLE
-   Rollover-Logfile -Logfile C:\temp\logfile.log
-#> 
-function Rollover-Logfile
+    Rollover-Logfile -Logfile "C:\Windows\Temp\logfile.log" -MaxFileSizeKB 2048
+#>
+Function Rollover-Logfile
 {
+#Validate path and write log or eventlog
 [CmdletBinding()]
 Param(
       #Path to test
       [parameter(Mandatory=$True)]
-      $Logfile,
+      [string]$Logfile,
       
       #max Size in KB
-      [parameter(Mandatory=$false)]
+      [parameter(Mandatory=$False)]
       [int]$MaxFileSizeKB = 1024
     )
 
     if (Test-Path $Logfile)
     {
-        $getLogfile = Get-Item $logFile
-        $logfileSize = $getLogfile.Length/1024
-        $newName = ($getLogfile.BaseName)
-        $newName += ".lo_"
-        $newLogFile = "$($getLogfile.Directory)\$newName"
-
-        if ($logfileSize -gt $MaxFileSizeKB)
+        $getLogfile = Get-Item $Logfile
+        if ($getLogfile.PSIsContainer)
         {
-            if (Test-Path $newLogFile)
+            # Just a folder. Skip actions
+        }
+        else 
+        {
+            $logfileSize = $getLogfile.Length/1024
+            $newName = "{0}.lo_" -f $getLogfile.BaseName
+            $newLogFile = "{0}\{1}" -f ($getLogfile.FullName | Split-Path -Parent), $newName
+
+            if ($logfileSize -gt $MaxFileSizeKB)
             {
-                #need to delete old file first
-                Remove-Item -Path $newLogFile -Force -ErrorAction SilentlyContinue
+                if(Test-Path $newLogFile)
+                {
+                    #need to delete old file first
+                    Remove-Item -Path $newLogFile -Force -ErrorAction SilentlyContinue
+                }
+                Rename-Item -Path ($getLogfile.FullName) -NewName $newName -Force -ErrorAction SilentlyContinue
             }
-            Rename-Item -Path $logFile -NewName $newName -Force -ErrorAction SilentlyContinue
         }
     }
 }
+#-----------------------------------------
 #endregion
